# AK: Replace years_to_days_since_epoch by near-instant function