Previously, the function `years_to_days_since_epoch` internally [looped over every single year](https://github.com/SerenityOS/serenity/blob/d26aabff0401690d6dd1332558b40f5cb64e4428/AK/Time.h#L69) between 1970 and the given argument, potentially taking many iterations. I wrote a benchmark that just calls `years_to_days_since_epoch` 100 times. It was very slow:
```
// old:
TestTime(48:48): Completed test 'years_to_days_since_epoch_points' in 59ms
TestTime(48:48): Completed test 'years_to_days_since_epoch_benchmark' in 2324ms
TestTime(48:48): Completed test 'years_to_days_since_epoch_span' in 0ms
```

The entire function (i.e. both loops) can be simply replaced by a single if, about 6 integer divisions, and a dozen additions. It's constant-time. (Not in the cryptographic sense; it just doesn't loop anymore.)
```
// new:
TestTime(49:49): Completed test 'years_to_days_since_epoch_points' in 0ms
TestTime(49:49): Completed test 'years_to_days_since_epoch_benchmark' in 0ms
TestTime(49:49): Completed test 'years_to_days_since_epoch_span' in 0ms
```

The speedup for real-life programs is probably negligible, because *usually* the number of iterations is low and the calls to that function rarely occur in a loop. However, the speedup is very much measurable for specifically-crafted programs like the above benchmark.