@BenWiederhake 2023-01-01
No flakes that I can see, I just did a *very* silly typo, and the tests detected that. Here's what the fix looks like (already squashed in the PR):
```diff
     int days = 365 * (year - 1970);
     int extra_leap_days = 0;
-    days += Detail::mod_zeros_in_range(begin_year, end_year, 4);
-    days -= Detail::mod_zeros_in_range(begin_year, end_year, 100);
-    days += Detail::mod_zeros_in_range(begin_year, end_year, 400);
+    extra_leap_days += Detail::mod_zeros_in_range(begin_year, end_year, 4);
+    extra_leap_days -= Detail::mod_zeros_in_range(begin_year, end_year, 100);
+    extra_leap_days += Detail::mod_zeros_in_range(begin_year, end_year, 400);
     return days + extra_leap_days * leap_sign;
```
D'oh!