@BenWiederhake 2023-01-01
Changes since first push:
- Completely avoided signed-integer overflow in all cases, in contrast to the old implementation (new commit)
- Fixed a silly typo introduced by that
- Improved the comment/explanation on the crucial `mod_zeros_in_range` function (thanks @nico!)
- Use a nicer type in the test (thanks @nico!)
- Remove a FIXME from LibJS, because it looks incorrect to me, and the old reasoning does not apply anymore.
- Use a more appropriate function in the Kernel