@BenWiederhake 2023-01-01
Changes since last push:
- Several attempts later, I think I finally got *actually* rid of integer overflow.

In other news, rebuild times are surprisingly high, and way too many things include AK/Time.h.