@BenWiederhake 2023-01-02
Changes since last push:
- Fixed wrong clang-format off/on tag (thanks @trflynn89!)
- Rebased for convenience