@BenWiederhake 2023-01-02
Changes since last push:
- Use `BENCHMARK_CASE` instead of `TEST_CASE` for benchmark (thanks @trflynn89!)
- Rebased for convenience

Output after the first commit (which introduces only the benchmark, not the speedup):
```
Running test 'years_to_days_since_epoch_points'.
Completed test 'years_to_days_since_epoch_points' in 6ms
Running benchmark 'years_to_days_since_epoch_benchmark'.
Completed benchmark 'years_to_days_since_epoch_benchmark' in 296ms
```

Output after the speedup:
```
Running test 'years_to_days_since_epoch_points'.
Completed test 'years_to_days_since_epoch_points' in 0ms
Running benchmark 'years_to_days_since_epoch_benchmark'.
Completed benchmark 'years_to_days_since_epoch_benchmark' in 0ms
```