@trflynn89 2023-01-02
@BenWiederhake While testing this out, I came across a year that has different results now than it did before (not due to overflow): `-999999` (test262 loves this year, the wine must've been good or something).

Before:
```
years_to_days_since_epoch(-999999) == -365961662
```

After:
```
years_to_days_since_epoch(-999999) == -365961663
```

It seems the former is correct based on some test262 expectation?