@BertalanD 2023-01-02
FWIW, I found another interesting implementation of this algorithm, maybe you could verify this one against that: https://howardhinnant.github.io/date_algorithms.html#days_from_civil