@trflynn89 2023-01-02
> Oh, and this old issue of mine might be somewhat relevant: #12729

I thought this all looked very familiar, for more discussion/test cases, see #13138