@BenWiederhake 2023-01-02
Turns out the logic is right, it's just that I tripped up on C++ division logic. I still can't wrap my head around the question why oh WHY C has integer division rounding toward zero. This makes everything so more complicated!

Specifically, `AK::Detail::mod_zeros_in_range(-3, 3, 4)` should be 1, but right now computes 2, although there aren't two zeros. I'm working on it, just wanted to give an update.