@BenWiederhake 2023-01-02
Changes since last push:
- Marked #12729 for auto-closure as suggested by @BertalanD.
- Fixed division logic and added test. Division-rounding-to-negative-inf is surprisingly non-trivial in C++. This resolves the issue of years_to_days_since_epoch(-999999). The PR now also contains a test for specifically that value.
- Converted the LibJS changes to a revert commit, as suggested by @trflynn89. Thanks!