@trflynn89 2023-01-02
> Marked LibC: localtime is very slow for large time_t values #12729 for auto-closure as suggested by @BertalanD.

That's not fixed here, is it? `time_to_tm` is still using the slow `for`-loop.