@BenWiederhake 2023-01-02
> `time_to_tm`

We seriously have two implementations of that? Oh god no.

Can I just replace the code duplication by a call to AK::Time, what do you think?