@trflynn89 2023-01-02
> Can I just replace the code duplication by a call to AK::Time, what do you think?

For sure, we are free to call into `AK::Time` from LibC.