@BenWiederhake 2023-01-02
> For sure, we are free to call into `AK::Time` from LibC.

Reading a bit more closely, time_to_tm is the inverse function, so it does "the opposite" of years_to_days_since_epoch. In fact, tm_to_time automatically got the speedup, because it just forwards to years_to_days_since_epoch.

*Theoretically*, one might multiply/divide by 365.2425, which always gives a result that is off by at most one. That would give an excellent initial guess, which can then be refined through years_to_days_since_epoch. However, I'm not going to implement that, especially not in this PR.